[epochX3] Apply again reverted change - Add {} and indent by two spaces xxx/FFX calls in auto CPPProcess.cc; same perf (Prepare to add the loop over ipagV/ievt)
After several tests, I decided that for the moment I would implement in auto the same strategy as in manual.
This is because it is the fastest, even if it looks quite ugly.
Some data buffers are prepared outide the loop and then used in side the loop.
All other alternatives result in a 10% performance hit on eemumu.
This can be reevaluated after completing epochX3.