[mch] in codegen, remove "wrows = %(sizew)d" which was giving 5 instead of 6 with multichannel (madgraph5/madgraph4gpu#473)
NB1: what is wrong is the wrows=5 in .mad code, not the wrows=6 in standalone code! but it is commented out anyway...

NB2: it is peculiar how two different sets of variables are used for mgOnGpuConfig.h and CPPProcess.h,
e.g. ncomb and nbhel for the numbers of helicities, which are computed using two different methods...
One could consider cleaning up and removing all these variables from CPPProcess.h for simplicity...